Don't copy auth states on receipt of a COMMIT message
Similarly to the previous commit, if we receive a COMMIT message, we
should just use it to start a new SIGMA protocol, rather than comparing
it against the old COMMIT message we may have still lying around.  This
completely removes the need for the otrl_auth_copy_on_commit function.