Test: Add new test-suite in tests/
This is a completely new test suite that adds unit and regression tests.
Still not everything is covered but it's at 76.3% line coverage and
87.5% function coverage. See the current status here.

	https://coverage.otr.im/

In order to run the tests, simply type "make check" from the top
directory of the repository.

The test suite uses libtap and adds both a shell and C library located
in tests/utils/tap.

Note that this test suite is not related at all to the tests located in
the directory test_suite. Future work should be done to either merge
them or remove the old one.

Tested with libgcrypt 1.5.4 and 1.6.2.

Signed-off-by: Julien Voisin <julien.voisin@dustri.org>
Signed-off-by: David Goulet <dgoulet@ev0ke.net>
Signed-off-by: Ian Goldberg <iang@cs.uwaterloo.ca>

Fixes #8